+++ server/dfr3-service/src/main/java/edu/illinois/ncsa/incore/service/dfr3/controllers/MappingController.java
@@ -11,6 +11,7 @@
 package edu.illinois.ncsa.incore.service.dfr3.controllers;
 
 import edu.illinois.ncsa.incore.common.AllocationConstants;
+import edu.illinois.ncsa.incore.common.auth.Authorizer;
 import edu.illinois.ncsa.incore.common.auth.IAuthorizer;
 import edu.illinois.ncsa.incore.common.auth.Privileges;
 import edu.illinois.ncsa.incore.common.dao.ISpaceRepository;
@@ -210,6 +211,7 @@ public class MappingController {
         }
 
         mappingSet.setCreator(username);
+        mappingSet.setOwner(username);
 
         String id = this.mappingDAO.saveMappingSet(mappingSet);
 
@@ -236,7 +238,8 @@ public class MappingController {
         Optional<MappingSet> mappingSet = this.mappingDAO.getMappingSetById(id);
 
         if (mappingSet.isPresent()) {
-            if (authorizer.canUserDeleteMember(username, id, spaceRepository.getAllSpaces(), groups)) {
+            Boolean isAdmin = Authorizer.getInstance().isUserAdmin(this.groups);
+            if (this.username.equals(mappingSet.get().getOwner()) || isAdmin) {
 //              remove id from spaces
                 List<Space> spaces = spaceRepository.getAllSpaces();
                 for (Space space : spaces) {